Added basic C++ project
The project can be compiled with CMake. It is recommended to run
these commands, starting from the root of the repository:

```bash
mkdir build/
cd build/
cmake ..
cmake --build .
```

The program will then be runnable by launching the `stmeltdown`
executable.

Currently, the program does nothing and exits with a code of 0.