SDL2 dependency and basic window + main loop
Requires the Simple DirectMedia Library library version >=2.0.0,
which can be found at https://libsdl.org/ or in your package
manager.

This program will display a window and will exit with code 0 when
the window is closed, or exit with code 1 if something went wrong
(e. g. SDL couldn't be initialized).