Textures with SDL_image
The game can now load and display images in the window.

This requires the SDL2_image library, which can be found at
https://www.libsdl.org/tmp/SDL_image/ or from your package
manager.

The `mk/` directory was added to include files necessary for
certain platforms or buildsystems. In this case, CMake needs a
search file to properly find SDL2_image.