Basic Doxyfile for Doxygen documentation
By default does not include the comments, for shortness and ease of
navigation. The comments can be generated in a side file with:

    doxygen -g Doxyfile.tmp

Also updated the .gitignore to ignore .bak files, generated when
updating the documentation. The Doxyfile should be updated regularly
(for new features) and after each modification (to ensure proper
formatting). The command to update a Doxyfile is:

    doxygen -s -u Doxyfile

The `-s` option prevents Doxygen from adding an explanatory paragraph
before each option. Please always use it to avoid cluttering the
source code with comments that don't apply to the project; users who
wish to have an easy access to comments can generate them using the
first comment described in this commit message.