CLI argument parsing stub
The game may now be launched with arguments from a terminal.

Currently supported options:

- `--help`

Will print a short text explaining the options, then exit successfully.

Example:

```
usage: stmeltdown [OPTIONS...]

Options:
  --help          Show this help text and exit
  --version       Show version info and exit
```

- `--version`

Will print the program's canonical name (`stmeltdown`), and its version
or "unknown" if the version was not compiled in (as of this commit, the
version will always be unknown).

Example:

```
stmeltdown v0.0.0
```