Version fetching from git tags
CMake will now grab the game's version based on the latest tag in
the repo. It uses `git describe --tags --always` to generate the
version string. It currently requires `git` to be in the `PATH`;
if not, the macro is not defined. The macro shouldn't be always
defined, and the code should not rely on its existence; read
further for the explanation.

From now on, all C++ files can access the game's version string
through the macro `STM_VERSION`. The GameManager already made use
of that macro (with an `#ifndef` guard), so it automatically gets
the version and manages it properly.

Similarly to what's done in GameManager, all code using this macro
should use a guard to prevent failing if the version string is
missing. A string may be missing because the command failed, or
because someone is compiling the game without using CMake.