Allow changing the data directory through CLI
Adds the `--data` CLI option, which changes the path in which to
look for assets.

This also required changing to code to take the data directory into
account.

Usage:

    ./stmeltdown --data path/to/data/

Assuming `path/to/data/` leads to a folder containing the various
data files.

If the option is omitted, it defaults to `../data/`, which, when
compiling the application within a build folder using the steps
descripted in `2a21a34`, allows an in-place invocation of the
freshly compiled binaries (with a working directory that corresponds
to the build folder) to find its data in the repo's source data
folder. That is, the defualt value allows running a fresh build
in-place.

In the future, more code may be added to also look in common
data folders, such as `/usr/{local/,}share/stmeltdown` on Unix,
or different paths on other systems.