Installation and packaging
This commit adds content necessary for the installation and packaging
of builds.

It currently supports packaging source builds only. Other formats,
such as .deb's, will come later.

In `CMakeLists.txt`, the project name has been changed to lowercase
because file names for packages re-use the project's name, including
capitalization.

In `mk/install/stmeltdown.desktop`, the icon name excludes the extension
to conform to the requirements for AppImages, which will be added soon.
Also, friendly reminder that `Version` refers to the version of the
file format, not the version of the game ;-)