Tile ghost on tilemap
A "ghost" of the currently selected tile will be displayed on the
tilemap at the position hovered by the mouse.

This also adds utility functions to change screen points to tilemap
points and vice versa.

It also fixes a discovered bug where a 0-height or 0-width tilemap
would push a new transform of the DrawingContext and never pop it.