PhysFS usage and support
Given that file I/O requires platform-dependant code for certain
operations (such as creating a directory), an external library is
added to handle those operations.

This also avoids propagating the data folder to all components that
need it, to the cost of global state. To ease the problem, the global
state management is kept alongside SDL.

The current method of handling paths is not resilient; a comment has
been left in `src/util/fs.cpp` to that effect. This will be fixed in
the future.

In addition, a few improvements have made their way in this commit,
such as the removal of needless includes (added by my wonderful IDE).