Run coverage even if tests fail
Given the preceeding commit, the test suite now exits with the proper
status code. However, the script that generates the coverage did not
take in consideration the fact that the test suite could fail, and
would abort the process if so. This commit fixes that problem by
instructing the script to ignore the exit code of the test suite.