Code quality and repository control tools
Some tools to check and ensure the proper quality of the repository
have been added.

Currently, it checks for:
- Respecting the 80 character limit in source code files
- Checking the copyright headers and the include guards
- Checking the Doxyfile for updates and formatting

More checks may be added in the future.

The scripts have been designed so a CI can simply invoke
`mk/tools/check_code_quality.sh` and the exit code of the
script will tell if there is any problem to fix. Information
is printed to the console as needed.