Library retrocompatility, GameManager refactor
The game now supports compiling with older versions of libraries (to a certain
degree). This should make it easier to port the game to various platforms, such
as certain Linux distros, smaller OSes with less support, or the Steam Linux
Runtime.

The GameManager class has also been refactored to split different tasks into
different functions. This is in preparation to splitting it into different
classes.

Fixes problems in the GameManager class:
- Dependencies (SDL/PhysFS) would be initialized even when they weren't needed
  (`--help` and `--version`).
- Dependencies wouldn't be unloaded properly if the game made an early exit
  (both above + `--test`).
- If the game launched and player normally, the return code would always be 0,
  regardless of whether or not an uncaught exception caused an unexpected exit.
- The return code of some dependency functions wasn't checked.

Also, more code is now attentive to the return code of functions that may fail.