Move loop logic away from main loop
This splits the main loop login into the loop per se and the code to execute
for a single loop.

This will make it easier to port the game to Emscripten, which uses a
different main loop mechanism. Additionally, it allows running frames one at a
time, which will be useful for a crash recovery mechanism in the future.