Move catch-all logic to separate function
Since the long try-catch that catches everything was repeated at many
places in the code (and will probably be repeated again several times
in the upcoming recovery mechanism), it has been moved to its own
function.

It is a template function, but does not need the same preprocessor
gymnastic as `Math::clamp()` in `src/util/math.*pp` does, since it
is a private function and will only be called within the same
compilation unit.