Add a crash recovery mechanism
Upon receiving and uncaught exception, the game will now try to re-launch
the game with as minimal losses as possible instead of simply closing the
game.

Currently, it tries to:
1. Restart the game without any change, in case the exception was a punctual
   problem.
2. Flush all caches, assets, and other non-essential data.
3. Remove the topmost scene, and resume from the parent scene.
4. Remove all scenes, and restart the game.
5. If all above fails, tell the user to manually restart the game, then crash.

In the future, more steps may be added, such as telling the scenes to make a
backup of their data and restarting the scenes.

Also includes some minor tweaking:
- Merged `launch_game()` and `main_loop()` since they did not seem to have any
  reason to be separate anymore.
- Moved initial scene to macro for ease of change and because it is now also
  used in the recovery function.