Parse CLI args before initing dependencies
Unlike what the message of `3e07365dd21615934e9edc73c6021dfd540560a8` says, the
CLI arguments were not parsed before initing dependencies. This fixes this
oversight.

This allows not only preventing `--help` and `--version` from not working if
SDL of PhysFS fail to load, but it also simplifies the two-step loading
(`init()` and `finish_setup()`) into a single function.

As such, the test suite is now responsible for loading/unloading dependencies
by itself. Currently, it does not test any code that requires SDL or PhysFS,
but this will be fixed in a later commit.

Also, the functions in GameManager were re-ordered to match the execution order
in the `run()` function.