API tests & output scanning
API tests have been added to the test suite. The naming template for test
functions has also been altered to accomodate for the new category.
(Specifically, the name of the category has been prepended to the function
names, separated from the rest by two underscores.)

To this effect, the LogScanner class has been added to the test system, to
ease the task for test function to scan for outputs given by the program.