Fix memory leak in EditorTilemap save/load
There was a memory leak that could happen when saving or loading a
tilemap. The SDL_Surface's were created the C way, which does not
play nicely with exceptions thrown notably by FS::get_rwops().

This changes the C pointers to C++ unique_ptrs, which ensures proper
destruction of data.