Actions in SceneController are optional
Functions in SceneController may or may not succeed, and return whether they do
succeed, or not.

Currently, the functions fail only if the required component is unavailable.

Having the ability to construct a SceneController without its dependencies
is mostly important to create mock objects within the test suite without
having to create a full GameManager/SceneManager.